requirements.txt
Flask
pymongo
python-dotenv
Werkzeug
PyJWT